fix(web): Add strict mode
closes: #9238
