+++ packages/web/src/generators/application/schema.json
@@ -69,6 +69,11 @@
       "enum": ["eslint", "none"],
       "default": "eslint"
     },
+    "strict": {
+      "type": "boolean",
+      "description": "Creates an application with strict mode and strict type checking.",
+      "default": true
+    },
     "skipFormat": {
       "description": "Skip formatting files",
       "type": "boolean",
